refactor(Inherited Code): Delete static variable and clean up BadLengthException class
[1] catch (BadLengthException e)
what it means is that in this line, 'e' is a copy of an exception thrown from checkUsername(). When you call copy ctor of a class, copy ctors of all of its member variables are called as well, so in this example, if a std::stringstream was part of a class, not a local variable of a constructor, a copy ctor of it would have to be called. But because it is marked as deleted, it can't be called, and that is why code can't be compiled.
(reference: https://en.cppreference.com/w/cpp/language/function#Deleted_functions)

[2] returning a pointer to an object (In this case, that is destroyed after function call)