feat(C++ Class Templates): Solve the 'C++ Class Templates' problem of c++ on HackerRank
This problem is not so good.
First, performance of environment that hackerrank offer is low.
So, submitting the code without 29 ~ 39line will result in a time out even if your code is correct.
Second, do not need specialization for 'concatenate' function. Because, its work is not difference to 'add' function in terms of performance. So, if you add '#define concatenate add', the code works without problem.